restructure into content folders
