+++ patrick/friends/src/misc/App.test.js
